+++ macroUtils/src/main/scala-2.11/TreeHelper.scala
@@ -1,0 +1,10 @@
+package io.github.cquiroz.utils
+
+import scala.reflect.macros.Context
+
+object TreeHelper {
+
+  def resetLong(c: Context)(v: c.Expr[Long]): c.Expr[Long] =
+    c.Expr[Long](v.tree.duplicate)
+
+}